+++ src/internal/connector/sharepoint/data_collections_test.go
@@ -92,7 +92,7 @@ func (suite *SharePointLibrariesSuite) TestUpdateCollections() {
 			newPaths := map[string]string{}
 			excluded := map[string]struct{}{}
 			c := onedrive.NewCollections(
-				graph.LargeItemClient(),
+				graph.HTTPClient(graph.NoTimeout()),
 				tenant,
 				site,
 				onedrive.SharePointSource,
